MSI-2283-Simplify-Source-Mapper-API. Remove CreateFromSourceInterface. Move all logic to the mapper. Update InStorePickup Module to depend on MapperInterface in code.
