+++ app/code/Magento/InventoryInStorePickup/Model/GetPickupLocationsAssignedToSalesChannel.php
@@ -10,8 +10,8 @@ namespace Magento\InventoryInStorePickup\Model;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\InventoryApi\Api\GetSourcesAssignedToStockOrderedByPriorityInterface;
-use Magento\InventoryInStorePickupApi\Model\Mapper;
 use Magento\InventoryInStorePickupApi\Api\GetPickupLocationsAssignedToSalesChannelInterface;
+use Magento\InventoryInStorePickupApi\Api\MapperInterface;
 use Magento\InventorySalesApi\Api\StockResolverInterface;
 
 /**
@@ -25,7 +25,7 @@ class GetPickupLocationsAssignedToSalesChannel implements GetPickupLocationsAssi
     private $getSourcesAssignedToStockOrderedByPriority;
 
     /**
-     * @var Mapper
+     * @var MapperInterface
      */
     private $mapper;
 
@@ -37,12 +37,12 @@ class GetPickupLocationsAssignedToSalesChannel implements GetPickupLocationsAssi
     /**
      * @param GetSourcesAssignedToStockOrderedByPriorityInterface $getSourcesAssignedToStockOrderedByPriority
      * @param StockResolverInterface $stockResolver
-     * @param Mapper $mapper
+     * @param MapperInterface $mapper
      */
     public function __construct(
         GetSourcesAssignedToStockOrderedByPriorityInterface $getSourcesAssignedToStockOrderedByPriority,
         StockResolverInterface $stockResolver,
-        Mapper $mapper
+        MapperInterface $mapper
     ) {
         $this->getSourcesAssignedToStockOrderedByPriority = $getSourcesAssignedToStockOrderedByPriority;
         $this->stockResolver = $stockResolver;
